+++ tests/testthat/test_homeranger.R
@@ -32,6 +32,8 @@ test_that("validate model run", {
     par = params,
     obs = system.file("extdata/Aspromonte_roedeer_traj.txt", package = "homeranger"),
     resolution = 25,
+    steps = 0,
+    runs = 0,
     optimization = FALSE,
     verbose = FALSE
   )
@@ -45,7 +47,7 @@ test_that("validate model run", {
 
 test_that("test optimizations", {
 
-  settings <- list(
+  params <- list(
     metric = hr_cost,
     control = list(
       sampler = "DEzs",
@@ -78,9 +80,10 @@ test_that("test optimizations", {
 
   # calibrate the model and optimize free parameters
   pars <- hr_fit(
-    drivers = raster_maps,
+    data = raster_maps,
     obs = system.file("extdata/Aspromonte_roedeer_traj_1196.txt", package = "homeranger"),
-    settings = settings,
+    resolution = 25,
+    par = params,
     parallel = FALSE
   )